move to util.py
