+++ common/util.py
